✨ feat: first naïve indexation
This is the first naïve indexation of the chain. This version index all
the accounts deployed on-chain. It stored all the account and the first signer
associated to them in the database. This implementation doesn't manage the
many-to-many relation between the accounts model and the signers model yet.
In addition to that, an minimalist ORM has been implemented for the accounts and
signers models.